RF_00.002.13 | Pipeline > Refactoring tests to POM
